refactor: attach keydown events on presentation component
