+++ frontend/src/pages/Presentation.vue
@@ -49,14 +49,15 @@ import SlideElementsPanel from '@/components/SlideElementsPanel.vue'
 import Slide from '@/components/Slide.vue'
 
 import { usePanAndZoom } from '@/utils/zoom'
+import { activeSlideIndex, name, presentation, activeSlideInFocus, position } from '@/stores/slide'
 import {
-	activeSlideIndex,
-	name,
-	presentation,
-	activeSlideElements,
-	activeSlideInFocus,
-} from '@/stores/slide'
-import { resetFocus } from '@/stores/element'
+	resetFocus,
+	currentFocusedIndex,
+	currentDataIndex,
+	duplicateElement,
+	addTextElement,
+} from '@/stores/element'
+import { duplicateSlide, changeSlide } from '@/stores/slideActions'
 import { saveChanges } from '@/stores/slideActions'
 
 let autosaveInterval = null
@@ -111,6 +112,67 @@ const startSlideShow = () => {
 	}
 }
 
+const updateElementPosition = (dx, dy) => {
+	if (!position.value) return
+	position.value = { left: position.value.left + dx, top: position.value.top + dy }
+}
+
+const handleArrowKeys = (key) => {
+	let dx = 0
+	let dy = 0
+
+	if (key == 'ArrowLeft') dx = -1
+	else if (key == 'ArrowRight') dx = 1
+	else if (key == 'ArrowUp') dy = -1
+	else if (key == 'ArrowDown') dy = 1
+
+	updateElementPosition(dx, dy)
+}
+
+const handleElementShortcuts = (e) => {
+	switch (e.key) {
+		case 'ArrowLeft':
+		case 'ArrowRight':
+		case 'ArrowUp':
+		case 'ArrowDown':
+			handleArrowKeys(e.key)
+			break
+		case 'Delete':
+		case 'Backspace':
+			deleteElement(e)
+			break
+		case 'd':
+			if (e.metaKey) duplicateElement(e)
+			break
+	}
+}
+
+const handleSlideShortcuts = (e) => {
+	switch (e.key) {
+		case 'ArrowUp':
+			changeSlide(activeSlideIndex.value - 1)
+			break
+		case 'ArrowDown':
+			changeSlide(activeSlideIndex.value + 1)
+			break
+		case 'Delete':
+		case 'Backspace':
+			deleteSlide()
+			break
+		case 'd':
+			if (e.metaKey) duplicateSlide(e)
+			break
+	}
+}
+
+const handleKeyDown = (e) => {
+	if (document.activeElement.tagName == 'INPUT' || currentFocusedIndex.value != null) return
+	if (e.key == 'Escape') return resetFocus()
+	if (e.key == 't') return addTextElement()
+
+	currentDataIndex.value ? handleElementShortcuts(e) : handleSlideShortcuts(e)
+}
+
 watch(
 	() => route.params.name,
 	async () => {
@@ -126,10 +188,12 @@ onMounted(() => {
 	zoom.targetElement.value = slideRef.value.targetRef
 	zoom.allowPanAndZoom.value = true
 	autosaveInterval = setInterval(saveChanges, 60000)
+	document.addEventListener('keydown', handleKeyDown)
 })
 
 onBeforeUnmount(() => {
 	zoom.allowPanAndZoom.value = false
 	clearInterval(autosaveInterval)
+	document.removeEventListener('keydown', handleKeyDown)
 })
 </script>
